base instructions for the GPT model almost ready
