+++ app/services/test_service.rb
@@ -1,0 +1,2 @@
+class TestService
+end
