+++ config/initializers/zeitwerk.rb
@@ -1,6 +1,6 @@
 # config/initializers/zeitwerk.rb
 Rails.autoloaders.each do |autoloader|
   autoloader.inflector.inflect(
-    'open_ai' => 'OpenAI'
+    'ai_tools' => 'AITools'
   )
 end
